+++ stories/components/typography/docs/typography.stories.tsx
@@ -3,12 +3,12 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import AdditionalInformationExample from './examples/additional-information';
 import BasicUsageExample from './examples/basic-usage';
 import FontSizeAndLineHeightExample from './examples/font-size-and-line-height';
-import FormatTextNestedListExample from './examples/formattext-nested-lists';
+import FormatTextNestedListExample, { defaultProps as FormatTextNestedListProps } from './examples/formattext-nested-lists';
 import HeadingMainExample from './examples/headings-main';
 import HeadingOtherExample from './examples/headings-other';
 import ListWithCustomBulletsExample from './examples/list-with-custom-bullets';
 import ListWithCustomContentExample from './examples/list-with-custom-content';
-import NativeTypographyTagExample from './examples/native-typography-tags';
+import NativeTypographyTagExample, { defaultProps as NativeTypographyTagProps } from './examples/native-typography-tags';
 import ParagraphMarginExample from './examples/paragraph-margins';
 import ParagraphsExample from './examples/paragraphs';
 import TextColorsExample from './examples/text-colors';
@@ -33,8 +33,14 @@ export const FontSizeAndLineHeight: Story = {
   render: FontSizeAndLineHeightExample,
 };
 
-export const FormatTextNestedList: Story = {
+export const FormatTextNestedList: StoryObj<typeof FormatTextNestedListProps> = {
   render: FormatTextNestedListExample,
+  argTypes: {
+    formatTags: {
+      control: { type: 'boolean' },
+    },
+  },
+  args: FormatTextNestedListProps,
 };
 
 export const HeadingMain: Story = {
@@ -53,8 +59,14 @@ export const ListWithCustomContent: Story = {
   render: ListWithCustomContentExample,
 };
 
-export const NativeTypographyTag: Story = {
+export const NativeTypographyTag: StoryObj<typeof NativeTypographyTagProps> = {
   render: NativeTypographyTagExample,
+  argTypes: {
+    formatTags: {
+      control: { type: 'boolean' },
+    },
+  },
+  args: NativeTypographyTagProps,
 };
 
 export const ParagraphMargins: Story = {